requirements.txt
amqp==5.1.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
bleach==6.0.0
celery==5.3.1
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-filter==23.2
django-redis==5.3.0
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-api-logger==1.1.14
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.3.1
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
